Breadth-First Update
Created the route class for recording routes.

created the findShortestPath method which uses a breadth-first search algorithm to find the shortest path between two Station objects. It initializes a queue and visited set, adds the starting station to the queue and visited set, and then iteratively explores each of the neighbors of the current station until it reaches the end station or exhausts all possible paths. If the end station is reached, it constructs a path using a previous map and returns a Route object that contains the path and the number of stops along the path. If no path is found, it returns null.

created the method bfsSearch which is called when the user wants to perform a breadth-first search to find the shortest path between two selected stations. It first checks if both start and end stations have been selected. If they have, it calls the Graph's findShortestPath method to find the shortest route between the selected start and end stations. If no path is found, it prints an error message. If a path is found, it prints the path and the number of stops.

slightly modified Station to implement the above two methods.