BFS Completed
New methods created:

the createStationCircle method creates two circles to represent a station on the map. The inner circle is filled with a specified color and the outer ring is drawn with the same color and a slightly larger radius, forming a ring around the inner circle. The location of the circles is determined by calling the calculateActualCoordinates method to convert the station's x and y coordinates into actual coordinates on the ImageView. The inner circle is added to the mapPane and the outer ring is added to it with a transparent fill and a colored stroke of width 2.

drawStartStationCircle() and drawEndStationCircle() are two private methods that are called when the user selects a start or an end station. These methods remove any existing circles from the mapPane, and then create new circles and outer rings at the location of the selected station. The createStationCircle() method is used to create the circles and outer rings, and takes the Station, Circle, Circle, and Color as arguments. Finally, the circles and outer rings are added to the mapPane.

The method drawLineBetweenStations creates a line object between two given stations on the map. The method takes in the starting and ending stations, as well as a color for the line. It calculates the actual coordinates of the two stations on the image using the calculateActualCoordinates method, creates a line object between those coordinates, and sets its stroke and width properties. Finally, it adds the line object to the mapPane.

The method "drawShortestPath" takes a "Route" object, which represents the shortest path between two stations, as an input. It then removes any existing lines on the map and draws lines between each station on the shortest path, using the "drawLineBetweenStations" method. The color of the lines is set to magenta

Bugs:

For some reason Notting Hill Gate is not counting Queensway as an adjacent station, super minor but needs to be kept an eye on, if one could be like this more could be, could throw off algorithms big time, will look into fixing it at some point.