Various Updates
Added new methods:

The "handleMapMouseClicked" method handles mouse clicks on the map. It adjusts the coordinates based on the image scale and sets the start station if the left button is clicked or the end station if the right button is clicked. It updates the map display accordingly.

The "findNearestStation" method calculates the nearest station based on given coordinates (x, y). It initializes variables to store the nearest station and its distance. It iterates over all stations and calculates the distance between the given coordinates and each station's coordinates. If the calculated distance is smaller than the current nearest distance, it updates the nearest station and nearest distance. Finally, it returns the nearest station.

The "calculateDistance" method calculates the distance between two points (x1, y1) and (x2, y2) using the Euclidean distance formula. The calculated distance is then returned.

The "selectMenuItem" method selects a specific MenuItem within a MenuButton based on a given stationName. It iterates over the items in the MenuButton's list and compares the text of each item with the given stationName. If a match is found, it sets the MenuButton's text to the selected item's text, effectively selecting that item.

The "clearMap" method clears the map by removing all lines from the mapPane. It uses the removeIf method to remove all nodes that are instances of javafx.scene.shape.Line. Additionally, it clears the items in the routeOutput ListView by calling the clear method on it.

The "closeApplication" method closes the application by calling the static exit() method of the Platform class in JavaFX.

Included line changes in output.