Dijkstra's Algorithm Finished
Added some new methods, most importantly:

dijkstraSearch method in the main conrtoller by calling the findShortestPathDijkstra method finds the shortest path between two selected stations using Dijkstra's algorithm. It checks if both start and end stations are selected. If so, it calculates the shortest route and displays it on the map. It also generates output for the ListView, showing the station path and line changes. The appearance of the list items is customized based on their index.

The findShortestPathDijkstra method in the graph class uses Dijkstra's algorithm to find the shortest path between two stations by exploring neighboring stations and updating distances along the way.

addNeighbor method in the station class, we calculate the Euclidean distance between the current station and a neighboring station. Then, we add the neighboring station to the map of neighbors along with its distance. Additionally, we keep track of the distances in a separate list.

calculateDistanceTo method in the station class calculates the Euclidean distance between the current station and another station. It uses the coordinates of both stations to calculate the distance using the Euclidean distance formula: taking the square root of the sum of the squares of the differences in the x and y coordinates.

Also made some minor adjustments here and there for some bugfixing, most importantly fixed an issue where a few stations were not recognising their adjacent stations.